+++ sdk/spring/azure-spring-data-cosmos/CHANGELOG.md
@@ -7,7 +7,7 @@
 #### Breaking Changes
 
 #### Bugs Fixed
-* Fixed when IN statement in annotated query didn't work with Long values
+* Fixed when IN statement in annotated query didn't work with Long values - See [PR 36249](https://github.com/Azure/azure-sdk-for-java/pull/36249).
 
 #### Other Changes